vpn_server/openvpn-pki/requirements.txt
Flask==1.0.2
flask-restplus==0.11.0
Flask-Login==0.4.1
pyopenssl
Werkzeug==0.16.0